# [#148491] Improve performance of FacilitiesController#show (Part 2)