--- a/app/views/facilities/show.html.haml
+++ b/app/views/facilities/show.html.haml
@@ -18,11 +18,11 @@
       = f.submit class: ['btn', 'btn-primary']
     = render partial: 'product_list', locals: { products: @instruments, f: f, title_extra: daily_view_link }
     - ["Service", "TimedService", "Item", "Bundle"].each do |type|
-      = render partial: 'product_list', locals: { products: @non_instrument_products_by_type[type], f: f }
+      = render partial: 'product_list', locals: { products: @non_instrument_products_by_type.fetch(type, []), f: f }
     .button_row
       = f.submit class: ['btn', 'btn-primary']
 - else
   .product_list_container{class: @columns}
     = render partial: 'product_list', locals: { products: @instruments, title_extra: daily_view_link }
     - ["Service", "TimedService", "Item", "Bundle"].each do |type|
-      = render partial: 'product_list', locals: { products: @non_instrument_products_by_type[type], f: f }
+      = render partial: 'product_list', locals: { products: @non_instrument_products_by_type.fetch(type, []) }