# Release Notes

Improve performance of FacilitiesController#show (Part 2)

# Additional Context

## Rendering all non-instruments before

```
   (0.4ms)  SELECT COUNT(*) FROM `products` WHERE `products`.`type` IN ('Service') AND `products`.`facility_id` = 1 AND `products`.`is_archived` = 0 AND `products`.`is_hidden` = 0
  Service Load (0.3ms)  SELECT  `products`.* FROM `products` WHERE `products`.`type` IN ('Service') AND `products`.`facility_id` = 1 AND `products`.`is_archived` = 0 AND `products`.`is_hidden` = 0 ORDER BY lower(products.name) LIMIT 1
  Service Load (0.3ms)  SELECT `products`.* FROM `products` WHERE `products`.`type` IN ('Service') AND `products`.`facility_id` = 1 AND `products`.`is_archived` = 0 AND `products`.`is_hidden` = 0 ORDER BY lower(products.name)
  Rendered collection of facilities/_product.html.haml [2 times] (2.3ms)
  Rendered facilities/_product_list.html.haml (11.0ms)
   (0.4ms)  SELECT COUNT(*) FROM `products` WHERE `products`.`type` IN ('TimedService') AND `products`.`facility_id` = 1 AND `products`.`is_archived` = 0 AND `products`.`is_hidden` = 0
  Rendered facilities/_product_list.html.haml (1.0ms)
   (0.3ms)  SELECT COUNT(*) FROM `products` WHERE `products`.`type` IN ('Item') AND `products`.`facility_id` = 1 AND `products`.`is_archived` = 0 AND `products`.`is_hidden` = 0
  Item Load (0.3ms)  SELECT  `products`.* FROM `products` WHERE `products`.`type` IN ('Item') AND `products`.`facility_id` = 1 AND `products`.`is_archived` = 0 AND `products`.`is_hidden` = 0 ORDER BY lower(products.name) LIMIT 1
  Item Load (0.3ms)  SELECT `products`.* FROM `products` WHERE `products`.`type` IN ('Item') AND `products`.`facility_id` = 1 AND `products`.`is_archived` = 0 AND `products`.`is_hidden` = 0 ORDER BY lower(products.name)
  Rendered collection of facilities/_product.html.haml [1 times] (1.1ms)
  Rendered facilities/_product_list.html.haml (6.2ms)
  Rendered collection of facilities/_product.html.haml [1 times] (1.1ms)
  Rendered facilities/_product_list.html.haml (4.1ms)
```

## Rendering all non-instruments after

```
  Rendered collection of facilities/_product.html.haml [2 times] (1.9ms)
  Rendered facilities/_product_list.html.haml (6.9ms)
  Rendered facilities/_product_list.html.haml (0.0ms)
  Rendered collection of facilities/_product.html.haml [1 times] (1.2ms)
  Rendered facilities/_product_list.html.haml (4.2ms)
  Rendered collection of facilities/_product.html.haml [1 times] (1.2ms)
  Rendered facilities/_product_list.html.haml (4.3ms)
```

Skylight reports [the SQL queries to load from `products` to be a large part of the response time](https://www.skylight.io/app/applications/Oj0TmDuSN-3Y/1557320400/1d/endpoints/FacilitiesController%23show?responseType=html), so this should help improve performance by executing fewer queries against that table.